feat: :sparkles: config react router
- add react router - config routes - create pages home, result and 404
