+++ package.json
@@ -19,6 +19,7 @@
     "postcss": "^8.5.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-router": "^7.6.0",
     "react-select": "^5.10.1"
   },
   "devDependencies": {
