+++ src/pages/result.tsx
@@ -1,0 +1,8 @@
+export const Result = () => {
+
+  return (
+    <section>
+      Result
+    </section>
+  )
+}
